requirements.txt
backoff==1.0.7
Django==1.9.4
django-ace==1.0.2
django-admin-tools==0.7.2
django-angular-scaffold==0.12.1
django-enumfield==1.2.1
django-environ==0.4.0
django-experiments==1.1.6
django-extensions==1.6.1
django-jet==0.1.4
django-jsonfield==0.9.12
django-jsonfield2==15.6.8
django-kong-admin==0.4.0
django-mailgun==0.8.0
django-modeldict==1.4.1
django-modeldict-yplan==1.5.0
django-reversion==1.10.1
django-subdomains==2.0.4
django-tastypie==0.13.3
enum34==1.1.2
gargoyle==0.11.0
google-api-python-client==1.5.0
gunicorn==19.4.5
httplib2==0.9.2
jsonfield==1.0.3
mock==1.0.1
nexus==0.3.1
oauth2client==2.0.1
ordereddict==1.1
Pillow==3.1.1
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.0
python-kong==0.4.0
python-mimeparse==1.5.1
pytz==2015.7
redis==2.10.5
requests==2.7.0
rsa==3.3
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
